Bolt and Animation
The test scene has the bolt script if you want to play around with it. I never could get gravity to function properly. Also I added a new battle sprite and animations for the new model. Not sure how to add him to the battle scene with the way things are configured.